Refactor tokens to submodule
