+++ tekore/_auth/__init__.py
@@ -1,6 +1,7 @@
-from .expiring import Credentials, Token, AccessToken
+from .expiring import Credentials
 from .refreshing import RefreshingCredentials, RefreshingToken
 from .scope import scope, Scope
+from .token import Token, AccessToken
 from .util import (
     parse_code_from_url,
     refresh_user_token,
